% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-question.R
\name{get_question}
\alias{get_question}
\title{Get A Question by ID}
\usage{
get_question(id = get_recent_ids()[1])
}
\arguments{
\item{id}{The ID of the question, which can be found in the URL (default to
the most recent question on the \link{r} tag on stackoverflow).}
}
\value{
An object of class \code{SOQuestion}
}
\description{
Get A Question by ID
}
\examples{
q <- get_question(54028838)
}
