% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-recent-ids.R
\name{get_recent_ids}
\alias{get_recent_ids}
\title{Get Recent \link{R} Question IDs}
\usage{
get_recent_ids()
}
\value{
An integer vector of length 30
}
\description{
This function grabs the most recent question ids for the \link{r} tag on
stackoverflow. These values can then be passed along to \code{get_question}
to retrieve the question that corresponds to the ID.
}
\examples{
ids <- get_recent_ids()
}
