% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-code.R
\name{run_code}
\alias{run_code}
\title{Run A Question's Code}
\usage{
run_code(q)
}
\arguments{
\item{q}{An object of class \code{SOQuestion}}
}
\value{
results and errors, invisibly
}
\description{
This function takes all formatted code blocks and attempts to run them in the
global environment.
}
\examples{
q <- get_question(54028838)
run_code(q)
}
