% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view-question.R
\name{view_question}
\alias{view_question}
\title{View a Question}
\usage{
view_question(q)
}
\arguments{
\item{q}{An object of class \code{SOQuestion}}
}
\description{
This function opens up the stackoverflow question in the Viewer pane in
RStudio.
}
\examples{
q <- get_question(54028838)
view_question(q)
}
